+++ CHANGES.md
@@ -1,7 +1,7 @@
 # Version 07.00.alpha15
 ## Changes
 * Meet scripts can set image on town screen
-
+* Removed Dump Bodies option for sell/fire girls
 
 
 # Version 07.00.alpha14